+++ crates/matrix-sdk-ui/src/timeline/mod.rs
@@ -142,11 +142,8 @@ pub enum TimelineFocus {
     Event {
         target: OwnedEventId,
         num_context_events: u16,
-        /// Whether to hide in-thread replies from the live timeline.
-        ///
-        /// This should be set to true when the client can create
-        /// [`Self::Thread`]-focused timelines from the thread roots themselves.
-        hide_threaded_events: bool,
+        /// How to handle threaded events.
+        thread_mode: TimelineEventFocusThreadMode,
     },
 
     /// Focus on a specific thread
@@ -154,6 +151,31 @@ pub enum TimelineFocus {
 
     /// Only show pinned events.
     PinnedEvents { max_events_to_load: u16, max_concurrent_requests: u16 },
+}
+
+/// Options for controlling the behaviour of [`TimelineFocus::Event`]
+/// for threaded events.
+#[cfg_attr(feature = "uniffi", derive(uniffi::Enum))]
+#[derive(Clone, Debug, PartialEq)]
+pub enum TimelineEventFocusThreadMode {
+    /// Force the timeline into threaded mode. When the focused event is part of
+    /// a thread, the timeline will be focused on that thread's root. Otherwise,
+    /// the timeline will treat the target event itself as the thread root.
+    /// Threaded events will never be hidden.
+    ForceThread,
+    /// Automatically determine if the target event is
+    /// part of a thread or not. If the event is part of a thread, the timeline
+    /// will be filtered to on-thread events.
+    Automatic {
+        /// When the target event is not part of a thread, whether to
+        /// hide in-thread replies from the live timeline. Has no effect
+        /// when the target event is part of a thread.
+        ///
+        /// This should be set to true when the client can create
+        /// [`TimelineFocus::Thread`]-focused timelines from the thread roots
+        /// themselves and doesn't use the [`Self::ForceThread`] mode.
+        hide_threaded_events: bool,
+    },
 }
 
 impl TimelineFocus {